Move struct `pair` into header `pair.hpp`
