+++ include/meevax/kernel/error.hpp
@@ -17,6 +17,7 @@
 #ifndef INCLUDED_MEEVAX_KERNEL_ERROR_HPP
 #define INCLUDED_MEEVAX_KERNEL_ERROR_HPP
 
+#include <meevax/kernel/collector.hpp>
 #include <meevax/kernel/pair.hpp>
 #include <meevax/kernel/string.hpp>
